+++ tests/integration/conftest.py
@@ -30,7 +30,7 @@ def api_key():
     try:
         return os.environ["AIRTABLE_API_KEY"]
     except KeyError:
-        pytest.skip("integration test requires AIRTABLE_API_KEY env variable")
+        pytest.skip("integration test requires AIRTABLE_API_KEY")
 
 
 @pytest.fixture
@@ -56,6 +56,11 @@ def base_name():
 @pytest.fixture
 def table_name():
     return "TEST_TABLE"
+
+
+@pytest.fixture
+def table_id():
+    return "tblfbOcVkVnnKxurq"
 
 
 @pytest.fixture